backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.1
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==1.1.0
httpx==0.28.1 